BEKIER_LACAILLE/requirements.txt
argparse==1.4.0
tinytag==1.6.0
mutagen==1.46.0
Pillow==9.5.0
tkinter==8.5
